% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_scalar_validator.R
\name{prop_scalar_validator}
\alias{prop_scalar_validator}
\title{Validate scalar properties of S7 classes}
\usage{
prop_scalar_validator(type, allow_null = FALSE)
}
\arguments{
\item{type}{A character specifying the type of scalar property. Possible values are "character" and "boolean".}

\item{allow_null}{A logical indicating whether NULL is an allowed value for the property.
When this is used, should be accompanied by defining the property's \code{class} argument to the union of NULL and a targeted class,
and using NULL as the property's \code{default} value.}
}
\value{
A function that returns a validation message if the value does not pass the validation, NULL otherwise.
}
\description{
Validator function factory for scalar properties. Use it to create functions to be passed as validators for S7 properties.
}
\examples{
prop_scalar_validator("character")(5)
prop_scalar_validator("boolean")(TRUE)
}
