% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skellm-package.R
\docType{package}
\name{skellm-package}
\alias{skellm}
\alias{skellm-package}
\title{skellm: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/calderonsamuel/skellm}
  \item \url{https://calderonsamuel.github.io/skellm/}
  \item Report bugs at \url{https://github.com/calderonsamuel/skellm/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
