% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skellm_message.R
\name{skellm_message_history}
\alias{skellm_message_history}
\title{Create a message history}
\usage{
skellm_message_history(...)
}
\arguments{
\item{...}{Any number of messages you provide. Each one of them must be a \code{skellm_message} object.}
}
\value{
An object with class \code{skellm_message_history}.
}
\description{
This class wraps a list of messages.
}
\examples{
skellm_message_history(
  skellm_message(
    role = "user",
    content = "What is 2 + 2"
  )
)

}
