% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-generic-skellm_request_perform.R
\name{skellm_req_perform}
\alias{skellm_req_perform}
\title{Perform a request}
\usage{
skellm_req_perform(skeleton, ...)
}
\arguments{
\item{skeleton}{The skeleton}
}
\value{
A httr2_response object
}
\description{
This performs a request
}
