% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skellm_message.R
\name{skellm_message}
\alias{skellm_message}
\title{Create a message}
\usage{
skellm_message(
  role = class_missing,
  content = class_missing,
  name = class_missing
)
}
\value{
An object with class \code{skellm_message}
}
\description{
This class represents a standard message format.
}
\section{Slots}{

\describe{
\item{\code{role}}{The role of the message's author. Typically one of "user", "system" or "assistant". We don't enforce a closed list of roles to offer more freedom.}

\item{\code{content}}{The content of the message.}

\item{\code{name}}{(Optional) The name of the message's author. Useful to differentiate between agents that have the same role.}
}}

\examples{
skellm_message(role = "system", content = "You are a helpful assistant")
skellm_message(role = "user", content = "What is 2 + 2")
skellm_message(role = "user", content = "<Object documentation>", name = "docs")

}
